requirements.txt
fastapi==0.115.0
uvicorn==0.40.0
python-dotenv==1.2.1
python-multipart==0.0.21
streamlit==1.53.0
pandas==2.3.3
plotly==6.5.2
requests==2.32.5
ollama>=0.4.7
openpyxl==3.1.5
pdfplumber==0.11.9
jinja2==3.1.5
altair==6.0.0
matplotlib==3.10.0
seaborn==0.13.2
numpy==2.2.0
